backend/requirements.txt
asgiref==3.5.2
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
install==1.3.5
mysqlclient==2.1.1
Pillow==9.2.0
PyJWT==2.4.0
pytz==2022.2.1
rest-framework-generic-relations==2.1.0
sqlparse==0.4.2
tzdata==2022.2
whitenoise==6.2.0
